chore: naming refactor
Made a change to how each file is named within the project, to ensure that the best practices are maintained and we are closer to the Atomic Design principles.

This change is made in interim before the next naming change that is to be implemented to ensure that there is full consistency between all files. Additionally more changes would be needed to be made so that the naming conventions are set for a clean project.
