+++ src/app/components/organisms/latestSec/latestArticleSection.module.css
@@ -2,8 +2,6 @@
     width: 100%;
     display: flex;
     justify-content: space-between;
-
-    margin-top: 75px;
 }
 
 .header {
