+++ src/app/page.tsx
@@ -1,10 +1,10 @@
 import { Inter } from 'next/font/google';
 import styles from './page.module.css';
-import Navbar from '@components/molecules/navbar/navbar';
-import HeroSection from '@components/organisms/herosec/heroSection';
-import WhySection from '@components/organisms/whySec/whySection';
-import LatestSection from '@components/organisms/latestSec/latestArticlesSection';
-import Footer from '@components/molecules/footer/footer';
+import Navbar from '@components/molecules/navbar/NavbarComponent';
+import HeroSection from '@components/organisms/herosec/HeroSectionOrg';
+import WhySection from '@components/organisms/whySec/WhySectionOrg';
+import LatestSection from '@components/organisms/latestSec/LatestArticlesOrg';
+import Footer from '@components/molecules/footer/FooterComponent';
 
 const inter = Inter({ 
   subsets: ['latin'],
